Tidied up some templates. Fixed some responsive glitches. Fixed error with editPicks and with login redirect.
Fixed the collapsing menu on myPicks page and added the same thing to the groupPicks page. Made the account page more responsive. Made the editPicks page more responsive. Fixed the login redirect error. Fixed an error with editPicks that only grabbed the first digit of the team id of winner. Formated some template files to be more readable.